se resolvio problema con delete y se agregaron api products
