+++ package.json
@@ -17,8 +17,7 @@
     "supertest": "^6.1.3"
   },
   "scripts": {
-    "dep": "nodemon solucion_/index.js",
-    "dev": "nodemon src/index.js",
+    "dev": "nodemon solucion_/index.js",
     "start": "NODE_ENV=production node src/index.js",
     "test:e2e": "jest --forceExit --config ./e2e/jest-e2e.json"
   },
