+++ solucion_/index.js
@@ -1,5 +1,4 @@
 //DB
-const app = require("./app")
+// const app = require("./app")
 require('dotenv').config({ path: 'variables.env' })
-const express = require("express");
-const app = require('app')
+const app = require("./db/app")
